Convert product texts to STEP-markup during migration
This is a first pass at converting hybris' HTML product texts into what
we colloquially call STEP-markup: a proprietary markup language similar
to but not quite SGML.

We want to both preserve as much original hybris content as possible and
minimise the subsequent human effort to reformat text in STEP. In a test
migration of 53,000 products from April 2017 this handles the 32 most
used tags out of 38. The result has not been approved by any content
owners but it has passed a smoke test ensuring legibility in STEP. There
may be later requirements to formatting changes, in particular tables
and URL references.

We know the following tags to be dirty:

- description
- summary
- maintenance
- ProductFeature.Value

The list may not be exhaustive.

The existing integration test compares input and output 1:1 and
naturally fails. Dumb down the test instead of resolving the newly added
complexity of attribute value differences.

We use an ancient version of jsoup that's already on the classpath.
There is already another, older version, and we don't want to make the
picture even more muddied.