Newline uses backslash'n', BlockElements Blockified.
Flawless Victory! - \n is now used instead of <return/>.

This change-set reflects a maturation of our ideas of what
constitutes a block element. A block element is an element that
"stands on its own" so to say - Something that is quite happy
to have its own formatting unaffected by its surroundings.

Previously block elements were a loosely defined construct
sometimes requiring a linebreak sometimes not. This lead to an
unnecessary amount of edge cases and made it rather difficult
to predict what would happen. The primary realisation was to
treat block elements as blocks. Elements independent of their
surroundings. I did this by absolutely defining within code
what constitutes a block element and always default to adding
newlines if unsure.

Formatting is now generously applied and all edge cases are
purged at the final step. This places the exported xml in a
happy medium where if nothing else there is consistency.
(Maximum of 3 consecutive newlines, Block elements are
never preceded by more than 3 newlines etc.)

The most important justification for this set of changes
existing at all is that we must preseve the text as it
appears in Hybris as STEP is not descriptive enough for
our needs.